Move all the common model classes in the same module
This refactoring permits to have in a single place the classes
to be used by both Cluster an ImageBuilder modules (e.g. Param, Resources, Tags)

The models/common.py now includes:
* Param and DynamicParam moved from param.py
* Resource and related classes from cluster.py
* Tag class (in common between Cluster and ImageBuilder)
* Generic Validator class and validation related classes
