+++ cli/src/pcluster/validators/ebs_validators.py
@@ -15,9 +15,8 @@ from pcluster.config.validators import (
     EBS_VOLUME_TYPE_TO_IOPS_RATIO,
     EBS_VOLUME_TYPE_TO_VOLUME_SIZE_BOUNDS,
 )
-from pcluster.models.param import Param
+from pcluster.models.common import FailureLevel, Param, Validator
 from pcluster.utils import get_ebs_snapshot_info, get_partition
-from pcluster.validators.common import FailureLevel, Validator
 
 
 class EbsVolumeTypeSizeValidator(Validator):
