+++ cli/src/pcluster/validators/s3_validators.py
@@ -6,8 +6,7 @@ from urllib.request import urlopen
 from botocore.exceptions import ClientError
 
 from common.boto3.s3 import S3Client
-from pcluster.models.param import Param
-from pcluster.validators.common import FailureLevel, Validator
+from pcluster.models.common import FailureLevel, Param, Validator
 
 
 class UrlValidator(Validator):
