+++ cli/tests/pcluster/validators/test_ec2.py
@@ -11,7 +11,7 @@
 
 import pytest
 
-from pcluster.models.param import Param
+from pcluster.models.common import Param
 from pcluster.validators.ec2_validators import (
     BaseAMIValidator,
     InstanceTypeBaseAMICompatibleValidator,
