+++ cli/tests/pcluster/validators/test_s3.py
@@ -1,6 +1,6 @@
 import pytest
 
-from pcluster.models.param import Param
+from pcluster.models.common import Param
 from pcluster.validators.s3_validators import UrlValidator
 from tests.pcluster.validators.utils import assert_failure_messages
